requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
djangorestframework==3.13.1
numpy==1.23.1
pkg_resources==0.0.0
psycopg2==2.9.3
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
